// apps/ui/src/lib/api-client/models/OAuthAnalyticsTopClient.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Proxy Smart Backend
 * SMART on FHIR Proxy + Healthcare Administration API using Keycloak and Elysia
 *
 * The version of the OpenAPI document: 0.0.5-alpha.202604142113.a183073f
 *
 * NOTE: This class is auto generated by generate-ts-fetch-client.py
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';

/**
 * @export
 * @interface OAuthAnalyticsTopClient
 */
export interface OAuthAnalyticsTopClient {
    /**
     * OAuth client ID
     * @type {string}
     * @memberof OAuthAnalyticsTopClient
     */
    clientId: string;
    /**
     * Client display name
     * @type {string}
     * @memberof OAuthAnalyticsTopClient
     */
    clientName: string;
    /**
     * Number of requests
     * @type {number}
     * @memberof OAuthAnalyticsTopClient
     */
    count: number;
    /**
     * Success rate percentage
     * @type {number}
     * @memberof OAuthAnalyticsTopClient
     */
    successRate: number;
}

/**
 * Check if a given object implements the OAuthAnalyticsTopClient interface.
 */
export function instanceOfOAuthAnalyticsTopClient(value: object): value is OAuthAnalyticsTopClient {
    if (!('clientId' in value) || value['clientId'] === undefined) return false;
    if (!('count' in value) || value['count'] === undefined) return false;
    if (!('successRate' in value) || value['successRate'] === undefined) return false;
    if (!('clientName' in value) || value['clientName'] === undefined) return false;
    return true;
}

export function OAuthAnalyticsTopClientFromJSON(json: any): OAuthAnalyticsTopClient {
    return OAuthAnalyticsTopClientFromJSONTyped(json, false);
}

export function OAuthAnalyticsTopClientFromJSONTyped(json: any, ignoreDiscriminator: boolean): OAuthAnalyticsTopClient {
    if (json == null) {
        return json;
    }
    return {
        
        'clientId': json['clientId'],
        'clientName': json['clientName'],
        'count': json['count'],
        'successRate': json['successRate'],
    };
}

export function OAuthAnalyticsTopClientToJSON(json: any): OAuthAnalyticsTopClient {
    return OAuthAnalyticsTopClientToJSONTyped(json, false);
}

export function OAuthAnalyticsTopClientToJSONTyped(value?: OAuthAnalyticsTopClient | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'clientId': value['clientId'],
        'clientName': value['clientName'],
        'count': value['count'],
        'successRate': value['successRate'],
    };
}